backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
opencv-python==4.9.0.80
ultralytics==8.1.0
easyocr==1.7.1
numpy==1.26.3
Pillow==10.2.0
aiofiles==23.2.1
twilio==8.12.0
email-validator==2.1.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3